Favor YARD over TomDoc
In the State of the Platform meeting on April 29, 2015, we discussed
whether the styleguide's recommendations were still relevant and
desirable to the team. One item brought up was that we allow for a
variety of method/function documentation and should favor one choice.
This PR makes the stand that YARD is our preferred documentation scheme.

Relevant information about the comparison between TomDoc and YARD can be
found at http://gnuu.org/2010/05/12/whats-missing-from-tomdoc/.